Create README.md
The Oregon Coast is known for having sharks in the region, which poses a risk for surfers who like to experience the great opportunities for recreation that the breaks provides.
The goal of this project is to understand the inherent risk of being attacked by a shark on the Oregon Coast. Hopefully, this analysis can help Oregon surfers understand the risks before paddling out.